backend/requirements.txt
# FastAPI and Server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pydantic==2.5.3
pydantic-settings==2.1.0

# ML and Computer Vision
torch==2.1.2
torchvision==0.16.2
transformers==4.37.0
segment-anything==1.0
groundingdino-py==0.4.0
opencv-python==4.9.0.80
Pillow==10.2.0
numpy==1.26.3

# Database
sqlalchemy==2.0.25
alembic==1.13.1
asyncpg==0.29.0
psycopg2-binary==2.9.9

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
bcrypt==4.1.2

# Storage & File Handling
boto3==1.34.34
supabase==2.3.0

# Utilities
python-dateutil==2.8.2
aiofiles==23.2.1
httpx==0.26.0
tenacity==8.2.3

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
httpx==0.26.0

# Code Quality
black==24.1.1
flake8==7.0.0
mypy==1.8.0
isort==5.13.2

# Monitoring
prometheus-client==0.19.0
sentry-sdk==1.40.0
